dsgf

+++ scripts_bash/parse_exceptions.sh
@@ -2,7 +2,7 @@
 
 ##input: work_dir customer ip app emails files
 if [ $# -lt 6 ];then
-  echo "We need work_dir customer ip app emails files"
+  echo "We need work_dir app files"
   exit 1
 fi
 
@@ -11,9 +11,6 @@ HERE=$(cd $(dirname "$0"); pwd)
 ALL_APPLICATIONS=( $(ls $HERE/../scripts_tcl/app_logs_paths/ | sed s/\.tcl\// | awk '{printf "%s ",$0} END {print ""}') )
 declare -a FILES
 OUR_APP=""
-EMAILS=""
-CUSTOMER=""
-REMOTE_IP=""
 WORK_DIR="/tmp"
 POS=0
 err=1
@@ -25,25 +22,12 @@ for var in ${@}; do
   let POS=$POS+1
   if [ $POS = 1 ]; then
     WORKDIR=$var;
-    echo "Working directory is $WORKDIR"
   elif [ $POS = 2 ]; then
-    CUSTOMER=$var;
-    echo "Customer is $CUSTOMER"
-  elif [ $POS = 3 ]; then
-    REMOTE_IP=$var;
-    echo "Remote ip is $REMOTE_IP"
-  elif [ $POS = 4 ]; then
     OUR_APP=$var;
-    echo "App is $OUR_APP"
-  elif [ $POS = 5 ]; then
-    EMAILS=$var;
-    echo "Emails are $EMAILS"
   else
     FILES=( "${FILES[@]}" "$var" )
   fi
 done
-
-echo "Files are ${FILES[@]}"
 
 for app in ${ALL_APPLICATIONS[@]}; do
   if [ $app == $OUR_APP ];then
@@ -76,15 +60,20 @@ fi
 . $SCRIPT_NAME
 
 #TMP_DIR=$(mktemp --tmpdir=$WORKDIR -u)
-CRT="$WORKDIR/current"
-PRV="$WORKDIR/previous"
-DIFF="$WORKDIR/new_difference"
-mv $CRT $PRV
-$OUR_APP > $CRT
-diff $PRV $CRT | grep "^>"| sed s/^\>\ // > $DIFF
+CRT=$WORKDIR/$OUR_APP\_current
+PRV=$WORKDIR/$OUR_APP\_previous
+DIFF=$WORKDIR/$OUR_APP\_new_difference
+RESULT=$WORKDIR/attachements/$OUR_APP\_new_exceptions.zip
+mkdir -p "$WORKDIR/attachements"
+if [ -f $CRT -a -s $CRT ];then
+  mv $CRT $PRV
+  $OUR_APP > $CRT
+  diff $PRV $CRT | grep "^<"| sed s/^\<\ // > $DIFF
+else
+  $OUR_APP > $CRT
+  cp $CRT $DIFF
+fi
 
-zip -9 $WORKDIR/new_exceptions.zip $DIFF
-OUR_APP=""
-CUSTOMER=""
-REMOTE_IP=""
-#echo Message | mailx -s _Subject_ -a $WORKDIR/new_exceptions.zip $EMAILS
+if [ -f $DIFF -a -s $DIFF ];then
+  zip -9 $RESULT $DIFF
+fi
